Split anim code to separate classes and convert to ES6
