+++ src/animation/animation.js
