+++ src/animation/skeleton.js
